Add files via upload
I figured out that each macro sort of works independently when run. So I had to put the InputBox command within the All Stocks Analysis macro in order to use that dynamic variable within the worksheet activation function. Before, I was attempting to run a separate macro for the InputBox function and trying to get that to work concurrently with the All Stocks Analysis macro. It returned a '9' error code.